puppet

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,8 @@
 # Puppet manifest to kill a process named killmenow
 exec { 'killmenow':
   command     => 'pkill -f killmenow',
+  onlyif  => 'pgrep -f killmenow',
   refreshonly => true,
+  unless      => 'pgrep -f killmenow',
   path        => ['/usr/bin/pkill']
 }
